+++ vcell-core/src/main/java/cbit/vcell/field/FieldDataDBOperationResults.java
@@ -10,14 +10,11 @@
 
 package cbit.vcell.field;
 
-import org.vcell.restclient.model.FieldDataReference;
 import org.vcell.util.document.ExternalDataIdentifier;
 import org.vcell.util.document.KeyValue;
-import org.vcell.util.document.User;
 
 import java.io.Serializable;
 import java.util.*;
-import java.util.stream.Collectors;
 
 
 public class FieldDataDBOperationResults implements Serializable {
@@ -30,22 +27,6 @@ public class FieldDataDBOperationResults implements Serializable {
 	public HashMap<ExternalDataIdentifier, Vector<KeyValue>> extdataIDAndSimRefH;
 
 
-	public static FieldDataDBOperationResults fieldDataReferencesToDBResults(List<FieldDataReference> dto, User user){
-		FieldDataDBOperationResults fieldDataDBOperationResults = new FieldDataDBOperationResults();
-		ArrayList<ExternalDataIdentifier> externalDataIdentifiers = new ArrayList<>();
-		ArrayList<String> externalDataAnnotations = new ArrayList<>();
-		HashMap<ExternalDataIdentifier, Vector<KeyValue>> externalDataIDSimRefs = new HashMap<>();
-		for (FieldDataReference fieldDataReference : dto){
-			ExternalDataIdentifier externalDataIdentifier = ExternalDataIdentifier.dtoToExternalDataIdentifier(fieldDataReference.getExternalDataIdentifier());
-			externalDataIdentifiers.add(externalDataIdentifier);
-			externalDataAnnotations.add(fieldDataReference.getExternalDataAnnotation());
-			List<KeyValue> keyValues = fieldDataReference.getExternalDataIDSimRefs().stream().map(KeyValue::dtoToKeyValue).collect(Collectors.toList());
-			externalDataIDSimRefs.put(externalDataIdentifier, new Vector<>(keyValues));
-		}
-		fieldDataDBOperationResults.extDataIDArr = externalDataIdentifiers.toArray(new ExternalDataIdentifier[0]);
-		fieldDataDBOperationResults.extDataAnnotArr = externalDataAnnotations.toArray(new String[0]);
-		fieldDataDBOperationResults.extdataIDAndSimRefH = externalDataIDSimRefs;
-		return fieldDataDBOperationResults;
-	}
+
 
 }
